+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Adds `ptr_eq()` to `Data` trait
 - Adds `trim()`, `trim_start()`, `trim_end()` methods to `ImString`
 - Adds `nom` integration
 - Adds `pop()`, `into_bytes()`, `as_mut_str()`, `from_utf16()` and `from_utf16_lossy()` methods for ImString.